requirements.txt
gradio==5.23.1
azure.cosmos==4.9.0
openai==1.66.5
openai-agents==0.0.7
dotenv==0.9.9
azure.identity==1.21.0
asyncio==3.4.3